requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-autoslug-field==0.2.3
django-ckeditor==6.0.0
django-extensions==3.0.9
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
